requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-apscheduler==0.7.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
pillow==11.0.0
playwright==1.48.0
psycopg2-binary==2.9.9
pyee==12.0.0
PyMySQL==1.1.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
